fix size_t to u64, it fixs compatibility with android build

+++ src/codec/picture.rs
@@ -58,7 +58,7 @@ impl<'a> Picture<'a> {
 
     pub fn new(format: format::Pixel, width: u32, height: u32) -> Result<Self, Error> {
         unsafe {
-            let ptr = av_malloc(mem::size_of::<AVPicture>() as size_t) as *mut AVPicture;
+            let ptr = av_malloc(mem::size_of::<AVPicture>() as u64) as *mut AVPicture;
 
             match avpicture_alloc(ptr, format.into(), width as c_int, height as c_int) {
                 0 => Ok(Picture {
